requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27

redis>=5.0

google-auth>=2.0
google-auth-oauthlib>=1.2
google-api-python-client>=2.0
requests>=2.31

beautifulsoup4>=4.12